(refactor): Separate modules into folders
