--- src/main.types.ts
@@ -1,4 +1,0 @@
-export enum NumberOfAttempts {
-  WorkflowId = 3,
-  WorkflowRuns = 5
-}
